process: Unshare UTS namespace, and set hostname to 'localhost'
Use UTS namespace to override hostname when network-sandbox is enabled.
Set it to 'localhost' as that has a better chance of being present
in /etc/hosts.  This fixes tests in some packages that try to connect
to localhost via hostname obtained using gethostname(), e.g. docker-py,
and suffer resolution problems due to the system hostname not being
defined in /etc/hosts.

Closes: https://github.com/gentoo/portage/pull/539
Signed-off-by: Michał Górny <mgorny@gentoo.org>