Add runtime fallback for _optimized_copyfile().
In some situations, as described by this thread on Gentoo forums:

https://forums.gentoo.org/viewtopic-t-1088232-start-0.html

the build system's kernel interface is not compatible with the portage
_optimized_copyfile() implementation.

Check for this specific edge-case, where _optimized_copyfile() fails
with errno=22 (Invalid argument), invoke the fallback function and set a
flag to indicate that following invocations will also call the fallback
function.

Also print a warning message at noiselevel=2 when this happens.

In the case where errno=22 is caused by some other reason which the user
should know about, the fallback function should fail in the same (or
similar) way.

Signed-off-by: Rafael Kitover <rkitover@gmail.com>