Add content-hash distfiles layout (bug 756778)
The content-hash layout is identical to the filename-hash layout,
except for these three differences:

1) A content digest is used instead of a filename digest.

2) The final element of the path returned from the get_path method
corresponds to the complete content digest. The path is a function
of the content digest alone.

3) Because the path is a function of content digest alone, the
get_filenames implementation cannot derive distfiles names from
paths, so it instead yields DistfileName instances whose names are
equal to content digest values. The DistfileName documentation
discusses resulting implications.

Motivations to use the content-hash layout instead of the
filename-hash layout may include:

1) Since the file path is independent of the file name, file
name collisions cannot occur. This makes the content-hash
layout suitable for storage of multiple types of files (not
only gentoo distfiles). For example, it can be used to store
distfiles for multiple linux distros within the same tree,
with automatic deduplication based on content digest. This
layout can be used to store and distribute practically anything
(including binary packages for example).

2) Allows multiple revisions for the same distfiles name. An
existing distfile can be updated, and if a user still has an
older copy of an ebuild repository (or an overlay), then a user
can successfully fetch a desired revision of the distfile as
long as it has not been purged from the mirror.

3) File integrity data is integrated into the layout itself,
making it very simple to verify the integrity of any file that
it contains. The only tool required is an implementation of
the chosen hash algorithm.

Bug: https://bugs.gentoo.org/756778
Signed-off-by: Zac Medico <zmedico@gentoo.org>