setup.py: prefix repos.conf paths for venv
Prefix repos.conf location and sync-openpgp-key-path values for venv. There
is no intention to do this for normal EPREFIX. This behavior is intended
only when create_entry_points is True, so that if emerge --sync runs in a
venv, the operation will not accidentally modify paths outside of sys.prefix.

Signed-off-by: Zac Medico <zmedico@gentoo.org>