Atom: Decouple eapi and allow_repo
This commit changes the Atom constructor such that
allow_repo is only overwritten with the eapi default
if allow_repo is None. If allow_repo is not None, its
value is preserved.

There are currently two deprecated EAPIs which allowed for repo atoms
("4-python" and "5-progress"). The current consensus seems to be that
repo atoms are not wanted in other eapis, and should instead be
implemented via a profile format extension [1]. As a prerequisite,
the allow_repo parameter has to be decoupled from the eapi parameter
as done by this commit, since otherwise if an eapi other than
"4-python" and "5-progress" is passed, this would force allow_repo to
be False. This would prevent implementing a profile format extension
that passes allow_repo=True.

A helper function allow_profile_repo_deps has been added
to initialize allow_repo based on the eapi of a repo before
calling the Atom constructor. This helper function can
be extended later to also take a profile format extension
into account.

All calls to the Atom constructor that pass both allow_repo and eapi
have been adapted to use this helper function, with the following
exceptions:

   update.py, vartree.py, bintree.py call isvalidatom to validate the
   rhs of a "move" command in profile/update [2].
   Before this commit, when using eapi "4-python" or "5-progress",
   this check permitted the rhs to be a repo atom. However, this
   use case seems to be unsupported even for those eapis: The code
   substitutes the lhs with the rhs in the target atom, so if the
   target atom contains further modifiers such as slots or use flags,
   the the substitution would place the ::repo before these modifiers,
   resulting in an invalid atom that is rejected by other parts of the
   code. For this reason, these calls have not been adapted.

[1]: https://bugs.gentoo.org/651208
[2]: https://dev.gentoo.org/~ulm/pms/head/pms.html#x1-360004.4.4

Signed-off-by: Felix Bier <felix.bier@rohde-schwarz.com>
Signed-off-by: Zac Medico <zmedico@gentoo.org>