Implement PROPERTIES=test_network and ALLOW_TEST
The 'test_network' property can be used to indicate that the test phase
requires access to the Internet (but RESTRICT=test should still be
used).  If present, network-sandbox will be disabled throughout the test
phase.  This opens up the possibility of adding further 'test_*'
properties.

Additionally, ALLOW_TEST can be used to ignore RESTRICT=test in a subset
of packages.  When the value includes 'network' token, tests using
'test_network' property are reenabled.  When the value includes 'all'
token, all instances of RESTRICT=test are ignored.

Bug: https://bugs.gentoo.org/553696
Signed-off-by: Michał Górny <mgorny@gentoo.org>