multilib_category.py: recognize LoongArch ABIs
Add its ELF machine number constant as well. The Gentoo ARCH is "loong";
all six ABIs defined in the LoongArch ELF psABI spec [1] are supported.
The upstream binutils port currently implements an earlier version of the
spec [2]; the ILP32 ABIs are never fully implemented so far, and the
value for LP64D is coincidentally the same, so the code here stays
compatible.

This is preparatory work towards a port to the LoongArch architecture,
which is the new RISC architecture from Loongson Corporation after
switching away from MIPS.

[1]: https://github.com/loongson/LoongArch-Documentation/blob/23d53fe146a4/docs/LoongArch-ELF-ABI-EN.adoc
[2]: https://sourceware.org/git/?p=binutils-gdb.git;a=blob;f=include/elf/loongarch.h;h=b4d801ae9bd5a6c4562ca6b1cbae6e6d45d6c704;hb=HEAD

Signed-off-by: WANG Xuerui <git@xen0n.name>
Closes: https://github.com/gentoo/portage/pull/740
Signed-off-by: Sam James <sam@gentoo.org>