install-qa-check.d/60pkgconfig: check for not respecting libdir in pc files
It's not valid to reference lib64 when installing to /usr/lib where we
want 32-bit libraries.

We want to make sure that if we're installing a pkgconfig file for a 32-bit
variant (multilib), we make sure that the file references the right library:
it should have e.g. /usr/lib, not /usr/lib64, or consumers trying to use
the 32-bit library will try to link against a 64-bit library.

(We also cover the opposite case: /usr/lib64 pkgconfig files referencing
/usr/lib).

Bug: https://bugs.gentoo.org/729642
Signed-off-by: Sam James <sam@gentoo.org>