estrip: fix lockfile handling
The previous code would wait for the first link to an inode to be
processed, and would then allow multiple processes to process duplicate
links to the same inode. This behavior leads to a race condition in
save_elf_debug.

The new code ensures that each inode is accessed by a single process at
a time.

Closes: https://bugs.gentoo.org/823798
Signed-off-by: Mike Gilbert <floppym@gentoo.org>