portage.output: Replace darkblue colors with teal
The "darkblue" color is often barely visible on dark terminals which
makes reading emerge output really hard (I basically have to copy-paste
it a lot in order to be able to read it at all).  Replace it with teal
that does not seem to have any significant use in the output.

Signed-off-by: Michał Górny <mgorny@gentoo.org>