repoman: make file.size-fatal non-fatal for non-::gentoo (overlays)
In #gentoo-dev-help, we've had a nummber of overlay users ask about
splitting patches. This defeats the purpose of the check:

in ::gentoo, we really want people to upload them to a separate location
& download them in SRC_URI. I don't think it's necessary for overlay
users to do this as long as they're aware it's a possible issue.

This downgrades file.size-fatal to a warning for repositories
other than ::gentoo.

Signed-off-by: Sam James <sam@gentoo.org>